removed ignored files
